backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.1
supabase==2.3.4
httpx==0.26.0
python-jose[cryptography]==3.3.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
